Add Common.Repository.Dapper

+++ Src/Common.BL/IEntityId.cs
@@ -2,6 +2,6 @@
 {
     public interface IEntityId
     {
-         int EntityId { get; }
+         long EntityId { get; }
     }
 }
